requirements.txt
compel==2.3.1
diffusers==0.37.1
Pillow==12.2.0
torch==2.5.1+cu121
ttkbootstrap==1.20.1
peft==0.19.0